+++ Avespoir.Core/Modules/Events/LogEvents.cs
@@ -1,4 +1,5 @@
-﻿using Avespoir.Core.Modules.Logger;
+﻿using Avespoir.Core.Configs;
+using Avespoir.Core.Modules.Logger;
 using Discord;
 using Discord.WebSocket;
 using System;
@@ -11,22 +12,22 @@ namespace Avespoir.Core.Modules.Events {
 		internal static Task LogEvent(LogMessage BotLogMessage) {
 			switch (BotLogMessage.Severity) {
 				case LogSeverity.Critical:
-					Log.Critical(BotLogMessage.Message, BotLogMessage.Exception);
+					Log.Critical($"Normal Log: {BotLogMessage.Message}", BotLogMessage.Exception);
 					break;
 				case LogSeverity.Debug:
-					Log.Debug(BotLogMessage.Message, BotLogMessage.Exception);
+					Log.Debug($"Normal Log: {BotLogMessage.Message}", BotLogMessage.Exception);
 					break;
 				case LogSeverity.Verbose:
-					Log.Verbose(BotLogMessage.Message, BotLogMessage.Exception);
+					Log.Verbose($"Normal Log: {BotLogMessage.Message}", BotLogMessage.Exception);
 					break;
 				case LogSeverity.Error:
-					Log.Error(BotLogMessage.Message, BotLogMessage.Exception);
+					Log.Error($"Normal Log: {BotLogMessage.Message}", BotLogMessage.Exception);
 					break;
 				case LogSeverity.Warning:
-					Log.Warning(BotLogMessage.Message, BotLogMessage.Exception);
+					Log.Warning($"Normal Log: {BotLogMessage.Message}", BotLogMessage.Exception);
 					break;
 				case LogSeverity.Info:
-					Log.Info(BotLogMessage.Message, BotLogMessage.Exception);
+					Log.Info($"Normal Log: {BotLogMessage.Message}", BotLogMessage.Exception);
 					break;
 			}
 
@@ -48,12 +49,24 @@ namespace Avespoir.Core.Modules.Events {
 			return Task.CompletedTask;
 		}
 
-		internal static Task DisconnectedEvent(Exception Error, DiscordSocketClient Bot) {
+		internal static async Task DisconnectedEvent(Exception Error, DiscordSocketClient Bot) {
 			Log.Info($"ShardID: {Bot.ShardId} Disconnected!");
 			if (!(Error is null)) {
 				Log.Error("Disconnection due to error", Error);
+				if (!ReadyEvent.ExitCheck) {
+					Log.Info("Reconnecting...");
+					ReadyEvent.ExitCheck = true;
+					await Task.Delay(1000).ConfigureAwait(false);
+
+					await Client.Bot.StopAsync().ConfigureAwait(false);
+					await Client.Bot.LogoutAsync().ConfigureAwait(false);
+
+					ReadyEvent.ExitCheck = false;
+
+					await Client.Bot.LoginAsync(TokenType.Bot, ClientConfig.Token).ConfigureAwait(false);
+					await Client.Bot.StartAsync().ConfigureAwait(false);
+				}
 			}
-			return Task.CompletedTask;
 		}
 
 		internal static Task LatencyUpdated(int Ping1, int Ping2, DiscordSocketClient Bot) {